add github config file and update conditions

+++ root/main/tools/etc.py
@@ -1,5 +1,5 @@
 
-__all__ = ('DJANGO_CONFIG', 'DATABASE_CONFIG', 'CLOUDINARY_CONFIG', 'GOOGLE_AUTH_CONFIG')
+__all__ = ('DJANGO_CONFIG', 'DATABASE_CONFIG', 'CLOUDINARY_CONFIG', 'GOOGLE_AUTH_CONFIG', 'GITHUB_AUTH_CONFIG')
 
 from pathlib import Path
 import json
@@ -8,16 +8,16 @@ BASE_DIR = Path(__file__).parents[3]
 CONFIG_DIR = BASE_DIR / 'etc'
 CONFIG_DIR.mkdir(exist_ok=True)
 
-
+# File config (input)
 DJANGO_FILE_CONFIG = CONFIG_DIR / 'config.json'
 DATABASE_FILE_CONFIG = CONFIG_DIR / 'db_config.json'
 CLOUDINARY_FILE_CONFIG = CONFIG_DIR / 'cloudinary_config.json'
 GOOGLE_AUTH_FILE_CONFIG = CONFIG_DIR / 'google_config.json'
+GITHUB_AUTH_FILE_CONFIG = CONFIG_DIR / 'github_config.json'
 
-# Creations
+
+# Write default settings (write)
 if not DJANGO_FILE_CONFIG.exists():
-    print(f"Configure your {CONFIG_DIR}")
-
     with DJANGO_FILE_CONFIG.open(mode='w') as js_file:
         json.dump({"DEBUG": "True",
                    "SECRET_KEY": "django-insecure-_____________________to_change____________________",
@@ -25,16 +25,22 @@ if not DJANGO_FILE_CONFIG.exists():
                    "EMAIL_PASSWORD": "",
                    "ALLOWED_HOSTS": ('127.0.0.1',)}, js_file)
 
+
+if not DATABASE_FILE_CONFIG.exists():
     with DATABASE_FILE_CONFIG.open(mode='w') as js_file:
         json.dump({'default': {'ENGINE': 'django.db.backends.sqlite3',
                                'NAME': str(BASE_DIR / 'database.sqlite3'),
                                'TEST': {'NAME': ':memory:', "ENGINE": "django.db.backends.sqlite3"}}}, js_file)
 
+
+if not CLOUDINARY_FILE_CONFIG.exists():
     with CLOUDINARY_FILE_CONFIG.open(mode='w') as js_file:
         json.dump({'cloud_name': "",
                    'api_key': "",
                    'api_secret': ""}, js_file)
 
+
+if not GOOGLE_AUTH_FILE_CONFIG.exists():
     with GOOGLE_AUTH_FILE_CONFIG.open(mode='w') as js_file:
         json.dump({"web": {"client_id": "",
                            "project_id": "",
@@ -44,7 +50,13 @@ if not DJANGO_FILE_CONFIG.exists():
                            "client_secret": ""}}, js_file)
 
 
-# Loadings
+if not GITHUB_AUTH_FILE_CONFIG.exists():
+    with GITHUB_AUTH_FILE_CONFIG.open(mode='w') as js_file:
+        json.dump({"DEV": {"client_id": "", "client_secret": ""},
+                   "PROD": {"client_id": "", "client_secret": ""}}, js_file)
+
+
+# Loadings (output)
 with DJANGO_FILE_CONFIG.open(mode='r') as js_file:
     DJANGO_CONFIG = json.load(js_file)
 
@@ -57,8 +69,13 @@ with CLOUDINARY_FILE_CONFIG.open(mode='r') as js_file:
 with GOOGLE_AUTH_FILE_CONFIG.open(mode='r') as js_file:
     GOOGLE_AUTH_CONFIG = json.load(js_file)
 
+with GITHUB_AUTH_FILE_CONFIG.open(mode='r') as js_file:
+    GITHUB_AUTH_CONFIG = json.load(js_file)
+
+
 if __name__ == '__main__':
     print(DJANGO_CONFIG)
     print(DATABASE_CONFIG)
     print(CLOUDINARY_CONFIG)
     print(GOOGLE_AUTH_CONFIG)
+    print(GITHUB_AUTH_CONFIG)
